Biegunka

Description

Biegunka - configuration management library

Interpreters control

biegunka

Arguments

:: (Controls -> Controls)
User defined settings

-> Script Profiles ()
Script to interpret

-> Interpreter
Combined interpreters

-> IO ()

Common Interpreters Controls wrapper

data Controls

Common interpreters controls

root :: Lens' Controls FilePath

Root path for Source layer lens

appData :: Lens' Controls FilePath

Biegunka profile files path

pretty :: Lens' Controls Pretty

Pretty printing

data Pretty

Constructors

Colors

Plain

Interpreters

pretend :: Interpreter

Pretend interpreter

Doesn't do any IO, so you can't check if script will fail to do IO

But Pretend can show which changes would be maid if IO will run without errors

Prints execution log if asked

 main :: IO ()
 main = pretend $ do
   profile ...
   profile ...

pause :: Interpreter

Simple interpreter example that just waits user to press any key

execute :: (EE -> EE) -> Interpreter

Execute Interpreter

Biegunka workhorse. Does useful work: copy and links files, compiles stuff, anything else

Supports setting execution options via its first argument

It's generally advised to use pretend before execute: that way you can catch some bugs in your script before devastation is done.

verify :: Interpreter

Verification interpreter

Compares current filesystem layout and what script says it should be line by line. Outputs errors it find, otherwise prints OK. Is useful to check execution correctness.

Execution environment hooks

data EE

Execution environment. Denotes default failure reaction, templates used and more

data Priviledges

Priviledges control. Controls how to behave if started with sudo

Constructors

Drop
Drop priviledges

Preserve
Preserve priviledges

priviledges :: Lens' EE Priviledges

data React

Constructors

Ignorant

Abortive

react :: Lens' EE React

data Templates

Wrapper for templates to not have to specify t type on ExecutionState Existence of that wrapper is what made Default instance possible

Constructors

forall t . ToSElem t => Templates t

templates :: Lens' EE Templates

retries :: Lens' EE Int

All layers

sudo :: String -> Script sc () -> Script sc ()

Change effective user id for wrapped commands

reacting :: React -> Script sc () -> Script sc ()

Change reaction pattern for wrapped commands

chain :: Script sc () -> Script sc () -> Script sc ()

Chain tasks sequentially Connects two tasks which forces them to run sequentially one after another.

Note: redundant if Order is Sequential

(~>>) :: Script sc () -> Script sc () -> Script sc ()

Alias for chain

Profile layer

profile :: String -> Script Sources () -> Script Profiles ()

Configuration profile

Provides convenient sources grouping

 profile "mine" $ do
   git ...
   git ...
 profile "friend's" $ do
   svn ...

File layer

registerAt :: FilePath -> Script Actions ()

Links source to specified filepath

 git "https://example.com/repo.git" "git/repo" $
   registerAt "we/need/you/here"

Links the whole ${HOME}/git/repo to ${HOME}/we/need/you/here

copy :: FilePath -> FilePath -> Script Actions ()

Copies given file to specified filepath

 git "https://example.com/repo.git" "git/repo" $
   copy "you" "we/need/you/here"

Copies ${HOME}/git/repo/you to ${HOME}/we/need/you/here

link :: FilePath -> FilePath -> Script Actions ()

Links given file to specified filepath

 git "https://example.com/repo.git" "git/repo" $
   link "you" "we/need/you/here"

Links ${HOME}/git/repo/you to ${HOME}/we/need/you/here

substitute :: FilePath -> FilePath -> Script Actions ()

Substitutes $template.X$ templates in given file and writes result to specified filepath

 git "https://example.com/repo.git" "git/repo" $
   substitute "you.hs" "we/need/you/here"

Substitutes templates in ${HOME}/git/repo/you.hs with values from Settings.template and writes result to ${HOME}/we/need/you/here

shell :: String -> Script Actions ()

Executes shell command with default shell

 git "https://example.com/repo.git" "git/repo" $
   shell "echo -n hello"

Prints hello (without a newline)

Convenient type aliases

data Script sc a

Newtype used to provide better error messages for type errors in DSL

data Scope

Constructors

Actions

Sources

Profiles