Biegunka.Execute.Control

Description

Controlling execution

Execution facade type

type Execution s a = TaggedT s (StateT EC IO) a

Stateful IO actions execution tagged with environment

Execution thread state control

data EC

Execution thread state. Denotes current failure reaction, effective user id and more

Constructors

EC

Fields

_reactStack :: [React]

Saved reactions modificators. Topmost is active

_usersStack :: [String]

Saved user chaning modificators. Topmost is active

_retryCount :: Int

Performed retries for task

reactStack :: Lens' EC [React]

Saved reactions modificators. Topmost is active

usersStack :: Lens' EC [String]

Saved user chaning modificators. Topmost is active

retryCount :: Lens' EC Int

Performed retries for task

Execution environment

data EE a

Execution environment. Denotes default failure reaction, templates used and more

Constructors

EE

Fields

_priviledges :: Priviledges

What to do with priviledges if ran in sudo

_react :: React

How to react on failures

_templates :: Templates

Templates mapping

_retries :: Int

Maximum retries count

_controls :: Controls

General biegunka controls

_stm :: a

Execution cross-thread state

_mode :: Mode

Execution mode

data STM

Concurrent parts of EE

Constructors

STM

Fields

_work :: TQueue Work

Task queue

_sudoing :: TVar Bool

Whether sudoed operation is in progress.

_running :: TVar Bool

Whether any operation is in progress.

_repos :: TVar (Set String)

Already updated repositories

priviledges :: Lens' (EE a) Priviledges

What to do with priviledges if ran in sudo

react :: Lens' (EE a) React

How to react on failures

templates :: Lens' (EE a) Templates

Templates mapping

retries :: Lens' (EE a) Int

Maximum retries count

stm :: Lens (EE a) (EE b) a b

Execution cross-thread state

work :: Lens' STM (TQueue Work)

Task queue

running :: Lens' STM (TVar Bool)

Whether any operation is in progress.

sudoing :: Lens' STM (TVar Bool)

Whether sudoed operation is in progress.

controls :: Lens' (EE a) Controls

General biegunka controls

repos :: Lens' STM (TVar (Set String))

Already updated repositories

mode :: Lens' (EE a) Mode

Execution mode

initializeSTM :: EE () -> IO (EE STM)

Prepare Execution environment to stm transactions

Misc

data Templates

Wrapper for templates to not have to specify t type on ExecutionState Existence of that wrapper is what made Default instance possible

Constructors

forall t . ToSElem t => Templates t

data Priviledges

Priviledges control. Controls how to behave if started with sudo

Constructors

Drop
Drop priviledges

Preserve
Preserve priviledges

data Work

Workload

Constructors

Do Int (IO ())
Task to come and its id

Stop Int
Task with that id is done

data Mode

How to do execution

Constructors

Dry
Dry run mode

Real
Real run mode