Biegunka.Language

Description

Specifies user side and library side languages primitives

Documentation

data Scope

Constructors

Actions

Sources

Profiles

newtype Script sc a

Newtype used to provide better error messages for type errors in DSL

Constructors

Script

Fields

runScript :: Free (EL sc) a

lift :: EL sc a -> Script sc a

Lift DSL term to the Script

data EL sc a where

External language datatype. That's what user will construct with combinators from Biegunka

Constructors

EP :: P -> Script Sources () -> a -> EL Profiles a

ES :: S -> Script Actions () -> a -> EL Sources a

EA :: A -> a -> EL Actions a

EW :: W -> a -> EL sc a

data IL

Constructors

IA A Int Int String String

IS FilePath String (IO ()) String String

IP String

IW W

IT [IL]

data A

Actions scope data

Constructors

Link FilePath FilePath
Symbolic link

Copy FilePath FilePath
Verbatim copy

Template FilePath FilePath (forall t. ToSElem t => t -> String -> Text)
Copy with template substitutions

Shell FilePath String
Shell command

data S

Sources scope data

Constructors

Source

Fields

stype :: String

Source type

suri :: String

URI where source is located

spath :: FilePath

Where to emerge source on FS (relative to Biegunka root setting)

supdate :: FilePath -> IO ()

How to update source

newtype P

Profiles scope data

Constructors

Profile

Fields

pname :: String

name

data W

Constructors

User (Maybe String)

Reacting (Maybe React)

Chain

data React

Constructors

Ignorant

Abortive