Biegunka.Primitive

Description

Language primitives

Profile layer primitives

profile :: String -> Script Sources () -> Script Profiles ()

Provides convenient Sources grouping. Does not nest

Example usage:

 profile "dotfiles" $ do
   git "https://github.com/supki/.dotfiles" ...
   git "https://github.com/dmalikov/dotfiles"
 profile "experimental" $ do
   git "https://github.com/ekmett/lens" ...

Actions layer primitives

link :: FilePath -> FilePath -> Script Actions ()

Links given file to specified filepath

 git "https://example.com/source.git" "git/source" $
   link "some-file" "anywhere"

Links ~/git/source/some-file to ~/anywhere. (Assuming default settings.)

register :: FilePath -> Script Actions ()

Links source to specified filepath

 git "https://example.com/source.git" "git/source" $
   register "somewhere"

Links ~/git/source to ~/somewhere. (Assuming default settings.)

copy :: FilePath -> FilePath -> Script Actions ()

Copies given file to specified filepath

 git "https://example.com/source.git" "git/source" $
   copy "some-file" "anywhere"

Copies ~/git/source/some-file to ~/anywhere. (Assuming default settings.)

substitute :: FilePath -> FilePath -> Script Actions ()

Substitutes templates in HStringTemplate syntax in given file and writes result to specified filepath

 git "https://example.com/source.git" "git/source" $
   substitute "some-file.template" "anywhere"

Copies ~/git/source/some-file.template to ~/anywhere. (Assuming default settings.)

Substitutes templates in ~/anywhere with values from templates part of Controls

shell :: String -> Script Actions ()

Executes shell command with default shell

 git "https://example.com/source.git" "git/source" $
   shell "echo hello"

Prints "hello\n" to stdout

Wrappers

sudo :: String -> Script sc () -> Script sc ()

Change effective user id for wrapped commands

reacting :: React -> Script sc () -> Script sc ()

Change reaction pattern for wrapped commands

chain :: Script sc () -> Script sc () -> Script sc ()

Chain tasks sequentially Connects two tasks which forces them to run sequentially one after another.

(<~>) :: Script sc () -> Script sc () -> Script sc ()

Alias for chain