Biegunka.Source.Dummy

Description

Biegunka.Source.Dummy - example Source using 'directory-layout'

Documentation

dummy

Arguments

:: Layout
Layout to make

-> FilePath
Layout root (relative to user home directory)

-> Script Actions ()
What to do with layout files

-> Script Sources ()

Make specified layout and attack it with Files

You can do with dummy Source whatever you can do with real ones.but it is easier to reason about because you don't depend on internet connection, installed libraries and so on.

dummy_

Arguments

:: Layout
Layout to make

-> FilePath
Layout root (relative to user home directory)

-> Script Sources ()

Make specified layout at given path and do nothing

This may seem as a function without any purpose but it might be actually useful if you want just to download Source somewhere