Biegunka.Source.Git

Description

Biegunka.Source.Git - functions to work with git repositories as sources

Source layer

git :: String -> FilePath -> Script Files () -> Script Source ()

Clone repository from the given url to specified path and/or pull from master. Also executes attached script

 git "https://example.com/repository.git" "git/repository" $ do
   registerAt "some/not/so/long/path"
   link "important.file" ".config"

clone repository from https://example.com/repository.git to ${HOME}/git/repository

pull from master

link ${HOME}/git/repository to ${HOME}/some/not/so/long/path

link ${HOME}/git/repository/important.file to ${HOME}/.config

git_ :: String -> FilePath -> Script Source ()

Clone repository from the given url to specified path and/or pull from master

 git_ "https://example.com/repository.git" "git/repository"

clone repository from https://example.com/repository.git to ${HOME}/git/repository

pull from master