Biegunka.Source.Zip

Description

Biegunka.Source.Zip - functions to work with .zip archives as sources

Source layer

zip :: String -> FilePath -> Script Files -> Script Source

Download and extract zip archive from the given url to specified path. Also executes attached script

 zip "https://example.com/archive.zip" "git/archive" $ do
   registerAt "some/not/so/long/path"
   link "important.file" ".config"

download and extract archive from https://example.com/archive.zip to ${HOME}/git/archive

link ${HOME}/git/archive to ${HOME}/some/not/so/long/path

link ${HOME}/git/archive/important.file to ${HOME}/.config

zip_ :: String -> FilePath -> Script Source

Download and extract zip archive from the given url to specified path.

 zip_ "https://example.com/archive.zip" "git/archive"

download and extract archive from https://example.com/archive.zip to ${HOME}/git/archive