Biegunka

Description

Biegunka - configuration development library

Interpreters control

biegunka

Arguments

:: (Settings () -> Settings ())
User defined settings

-> Interpreter
Combined interpreters

-> Script Profiles ()
Script to interpret

-> IO ()

Common Interpreters Controls wrapper

data Settings a

Settings common for all interpreters and also specific for this one

root :: Lens' (Settings a) FilePath

Root path for Source layer

appData :: Lens' (Settings a) FilePath

Biegunka profile files

colors :: Lens' (Settings a) ColorScheme

Pretty printing

data ColorScheme

Colors used in logger

Constructors

ColorScheme

Fields

_actionColor :: Doc -> Doc

_sourceColor :: Doc -> Doc

_srcColor :: Doc -> Doc

_dstColor :: Doc -> Doc

_errorColor :: Doc -> Doc

_retryColor :: Doc -> Doc

actionColor :: Lens' ColorScheme (Doc -> Doc)

Action color

sourceColor :: Lens' ColorScheme (Doc -> Doc)

Source color

srcColor :: Lens' ColorScheme (Doc -> Doc)

Src color

dstColor :: Lens' ColorScheme (Doc -> Doc)

Dst color

errorColor :: Lens' ColorScheme (Doc -> Doc)

Error color

retryColor :: Lens' ColorScheme (Doc -> Doc)

Retry color

noColors :: ColorScheme

Disable colors

Interpreters

pause :: Interpreter

Interpreter that just waits user to press any key

confirm :: Interpreter

Interpreter that awaits user confirmation

dryRun :: Interpreter

Dru run interpreter

run :: (Execution -> Execution) -> Interpreter

Real run interpreter

check :: Interpreter

Check interpreter

pretend :: Interpreter

Deprecated: Please, use dryRun

Dru run interpreter

execute :: (Execution -> Execution) -> Interpreter

Deprecated: Please, use run

Real run interpreter

verify :: Interpreter

Deprecated: Please, use check

Check interpreter

Execution environment hooks

data Execution

Executor environment. Denotes default failure reaction, templates used and more

data Priviledges

Priviledges control. Controls how to behave if started with sudo

Constructors

Drop
Drop priviledges

Preserve
Preserve priviledges

priviledges :: Lens' Execution Priviledges

What to do with priviledges if ran in sudo

data React

Failure reaction

Constructors

Ignorant

Abortive

Retry

react :: Lens' Execution React

How to react on failures

data Templates

Wrapper for templates to not have to specify t type on ExecutionState Existence of that wrapper is what made Default instance possible

Constructors

forall t . ToSElem t => Templates t

templates :: Lens' Execution Templates

Templates mapping

retries :: Lens' Execution Int

Maximum retries count

Profiles layer primitives

profile :: String -> Script Sources () -> Script Profiles ()

Provides convenient Sources grouping. Does not nest

Information about sources and files related to a particular profile profile could be found in ~/.biegunka/profile. (Assuming default settings.)

Example usage:

 profile "dotfiles" $ do
   git "https://github.com/supki/.dotfiles" ...
   git "https://github.com/dmalikov/dotfiles"
 profile "experimental" $ do
   git "https://github.com/ekmett/lens" ...

Sources layer primitives

class Sourceable s where

Common Sources structure

Methods

actions :: Lens' s (Script Actions ())

Actions to run after source update

(==>) :: String -> FilePath -> s -> Script Sources ()

Actions layer primitives

link :: FilePath -> FilePath -> Script Actions ()

Links given file to specified filepath

 git "https://example.com/source.git" "git/source" $
   link "some-file" "anywhere"

Links ~/git/source/some-file to ~/anywhere. (Assuming default settings.)

register :: FilePath -> Script Actions ()

Links source to specified filepath

 git "https://example.com/source.git" "git/source" $
   register "somewhere"

Links ~/git/source to ~/somewhere. (Assuming default settings.)

copy :: FilePath -> FilePath -> Script Actions ()

Copies given file to specified filepath

 git "https://example.com/source.git" "git/source" $
   copy "some-file" "anywhere"

Copies ~/git/source/some-file to ~/anywhere. (Assuming default settings.)

substitute :: FilePath -> FilePath -> Script Actions ()

Substitutes templates in HStringTemplate syntax in given file and writes result to specified filepath

 git "https://example.com/source.git" "git/source" $
   substitute "some-file.template" "anywhere"

Copies ~/git/source/some-file.template to ~/anywhere. (Assuming default settings.)

Substitutes templates in ~/anywhere with values from templates part of Controls

shell :: String -> Script Actions ()

Executes shell command with default shell

 git "https://example.com/source.git" "git/source" $
   shell "echo hello"

Prints "hello\n" to stdout

raw :: FilePath -> [String] -> Script Actions ()

Executes raw command

 git "https://example.com/source.git" "git/source" $
   raw "/bin/echo" ["-n", "hello"]

Prints "hello" to stdout

Modifiers

sudo :: String -> Script s () -> Script s ()

Change effective user id for wrapped commands

reacting :: React -> Script s () -> Script s ()

Change reaction pattern for wrapped commands

chain :: Script s () -> Script s () -> Script s ()

Chain tasks sequentially Connects two tasks which forces them to run sequentially one after another.

(<~>) :: Script s () -> Script s () -> Script s ()

Infix alias for chain

Types

data Script s a

Newtype used to provide better error messages for type errors in DSL

data Scope

Language terms scopes [kind]

Constructors

Actions

Sources

Profiles

TH

makeOptionsParser :: Name -> Q [Dec]

Make parser for biegunka and environment options

The following options become available:

--safe-run (default)

--run

--dry-run

--check

--full

Also one option for each environment (the --lowercased environment name)

The usage is trivial:

 data Environments = X220 | T510

 makeOptionsParser ''Environment

 main :: IO ()
 main = do
   (environment, runBiegunka) <- optionsParser
   ...